update scripts to generate arrays for filters
